allauthdemo/requirements.txt
django
django-allauth